feat: add a session variable to separate sink
