+++ src/frontend/planner_test/tests/testdata/output/sink.yaml
@@ -262,3 +262,64 @@
   explain_output: |
     StreamSink { type: upsert, columns: [v1, v2, v3, v4, t1._row_id(hidden)], downstream_pk: [t1.v1] }
     └─StreamTableScan { table: t1, columns: [v1, v2, v3, v4, _row_id] }
+- id: create_upsert_jdbc_sink_with_downstream_pk1_separate_sink
+  sql: |
+    set streaming_separate_sink to true;
+    create table t1 (v1 int, v2 double precision, v3 varchar, v4 bigint, v5 decimal, primary key (v3,v4));
+    explain create sink s1_mysql as select v1, v2, v3, v5 from t1 WITH (
+      connector='jdbc',
+      primary_key='v1,v2',
+      jdbc.url='jdbc:mysql://127.0.0.1:8306/mydb?user=root',
+      table.name='t1sink',
+      type='upsert');
+  explain_output: |
+    StreamSink { type: upsert, columns: [v1, v2, v3, v5, t1.v4(hidden)], downstream_pk: [t1.v1, t1.v2] }
+    └─StreamExchange { dist: HashShard(t1.v1, t1.v2) }
+      └─StreamTableScan { table: t1, columns: [v1, v2, v3, v5, v4] }
+- id: create_mock_iceberg_sink_upsert_with_range_partition_separate_sink
+  sql: |
+    set streaming_separate_sink to true;
+    create table t1 (v1 date, v2 timestamp, v3 timestamp with time zone, v4 timestamp);
+    explain create sink s1 as select v1 as v1, v2 as v2, v3 as v3, v4 as v4 from t1 WITH (
+      connector = 'iceberg',
+      type = 'upsert',
+      catalog.type = 'mock',
+      catalog.name = 'demo',
+      database.name = 'demo_db',
+      table.name = 'range_table',
+      warehouse.path = 's3://icebergdata/demo',
+      s3.endpoint = 'http://127.0.0.1:9301',
+      s3.region = 'us-east-1',
+      s3.access.key = 'hummockadmin',
+      s3.secret.key = 'hummockadmin',
+      primary_key = 'v1'
+    );
+  explain_output: |
+    StreamSink { type: upsert, columns: [v1, v2, v3, v4, t1._row_id(hidden)], downstream_pk: [t1.v1] }
+    └─StreamExchange [no_shuffle] { dist: UpstreamHashShard(t1._row_id) }
+      └─StreamTableScan { table: t1, columns: [v1, v2, v3, v4, _row_id] }
+- id: separate_aggregation_and_sink
+  sql: "set streaming_separate_sink to true;\ncreate table t (v1 date, v2 timestamp, v3 timestamp with time zone, v4 timestamp);\ncreate table sink_table (v1 date, v2 bigint);\nexplain create sink s into sink_table as select v1, count(*) from t group by v1 \nwith (type = 'append-only', force_append_only = 'true');\n"
+  explain_output: |
+    StreamProject { exprs: [t.v1, count, null:Serial] }
+    └─StreamSink { type: append-only, columns: [v1, count] }
+      └─StreamExchange [no_shuffle] { dist: HashShard(t.v1) }
+        └─StreamHashAgg { group_key: [t.v1], aggs: [count] }
+          └─StreamExchange { dist: HashShard(t.v1) }
+            └─StreamTableScan { table: t, columns: [v1, _row_id] }
+- id: separate_join_and_sink
+  sql: |
+    set streaming_separate_sink to true;
+    create table t (v1 date, v2 timestamp, v3 timestamp with time zone, v4 timestamp);
+    create table sink_table (v1 date, v2 timestamp);
+    explain create sink s into sink_table as select t1.v1, t2.v2 from t t1 join t t2 on t1.v1 = t2.v1
+    with (type = 'append-only', force_append_only = 'true');
+  explain_output: |
+    StreamProject { exprs: [t.v1, t.v2, null:Serial] }
+    └─StreamSink { type: append-only, columns: [v1, v2, t._row_id(hidden), t._row_id#1(hidden)] }
+      └─StreamExchange [no_shuffle] { dist: HashShard(t.v1) }
+        └─StreamHashJoin { type: Inner, predicate: t.v1 = t.v1 }
+          ├─StreamExchange { dist: HashShard(t.v1) }
+          │ └─StreamTableScan { table: t, columns: [v1, _row_id] }
+          └─StreamExchange { dist: HashShard(t.v1) }
+            └─StreamTableScan { table: t, columns: [v1, v2, _row_id] }
