+++ src/frontend/src/optimizer/rule/mod.rs
@@ -180,7 +180,7 @@ mod stream;
 pub use stream::bushy_tree_join_ordering_rule::*;
 pub use stream::filter_with_now_to_join_rule::*;
 pub use stream::generate_series_with_now_rule::*;
-pub use stream::separate_consecutive_join::*;
+pub use stream::separate_consecutive_join_rule::*;
 pub use stream::split_now_and_rule::*;
 pub use stream::split_now_or_rule::*;
 pub use stream::stream_project_merge_rule::*;
