+++ src/frontend/src/optimizer/rule/stream/mod.rs
@@ -15,7 +15,7 @@
 pub(crate) mod bushy_tree_join_ordering_rule;
 pub(crate) mod filter_with_now_to_join_rule;
 pub(crate) mod generate_series_with_now_rule;
-pub(crate) mod separate_consecutive_join;
+pub(crate) mod separate_consecutive_join_rule;
 pub(crate) mod split_now_and_rule;
 pub(crate) mod split_now_or_rule;
 pub(crate) mod stream_project_merge_rule;
