+++ src/frontend/src/optimizer/rule/stream/separate_consecutive_join_rule.rs
